Refactor choices, moving them to separate files
